Prevent collisions on first simulation tick().
Previously, if a large number of Bodies were generated at the beginning
of the simulation, collisions would occur on the first simulation iteration.
This caused colliding Bodies to have a larger initial velocity than those
that were not part of a collision.

To fix this, we run 100 iterations of collide() prior to the first tick().  This
seems to fix the issue but it would be nice to figure out the minimum
number of collide() iterations that we'll need to prevent this issue.

As it stands, 1 iterations is not always enough.

Signed-off-by: suicidepills <atom.mac@gmail.com>